+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [Parse command line arguments](https://github.com/amp-lang/ampc/issues/11)
 - [Disable returning types from functions entirely](https://github.com/amp-lang/ampc/issues/10)
 - [Ensure `return` statement has value](https://github.com/amp-lang/ampc/issues/12)
+- [Implement indirect functions as values in Cranelift IR](https://github.com/amp-lang/ampc/issues/7)
 
 ### [`snapshot-05-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-05-06-2023-a)
 - [Report diagnostics in `Type::from_ast`](https://github.com/amp-lang/ampc/issues/5) (report invalid types)